Top-deck crash fix
