+++ routes/socket/game/election.js
@@ -1,258 +1,8 @@
 const { sendInProgressGameUpdate, sendInProgressModChatUpdate } = require('../util');
-const { startElection, shufflePolicies, failedElection } = require('./common');
-const { sendGameList } = require('../user-requests');
-const {
-	specialElection,
-	policyPeek,
-	investigateLoyalty,
-	executePlayer,
-	selectPolicies,
-	selectPlayerToExecute,
-	selectPartyMembershipInvestigate,
-	selectSpecialElection
-} = require('./policy-powers');
+const { startElection, shufflePolicies, failedElection, enactPolicy } = require('./common');
 const { completeGame } = require('./end-game');
 const _ = require('lodash');
 const { makeReport } = require('../report.js');
-
-const powerMapping = {
-	investigate: [investigateLoyalty, 'The president must investigate the party membership of another player.'],
-	deckpeek: [policyPeek, 'The president must examine the top 3 policies.'],
-	election: [specialElection, 'The president must select a player for a special election.'],
-	bullet: [executePlayer, 'The president must select a player for execution.']
-};
-
-const presidentPowers = [
-	{
-		0: null,
-		1: null,
-		2: powerMapping.deckpeek,
-		3: powerMapping.bullet,
-		4: powerMapping.bullet
-	},
-	{
-		0: null,
-		1: powerMapping.investigate,
-		2: powerMapping.election,
-		3: powerMapping.bullet,
-		4: powerMapping.bullet
-	},
-	{
-		0: powerMapping.investigate,
-		1: powerMapping.investigate,
-		2: powerMapping.election,
-		3: powerMapping.bullet,
-		4: powerMapping.bullet
-	}
-];
-
-/**
- * @param {object} game - game to act on.
- * @param {string} team - name of team that is enacting policy.
- */
-const enactPolicy = (game, team) => {
-	const index = game.trackState.enactedPolicies.length;
-	const { experiencedMode } = game.general;
-
-	if (game.private.lock.selectChancellor) {
-		game.private.lock.selectChancellor = false;
-	}
-
-	if (game.private.lock.selectChancellorVoteOnVeto) {
-		game.private.lock.selectChancellorVoteOnVeto = false;
-	}
-
-	if (game.private.lock.selectChancellorPolicy) {
-		game.private.lock.selectChancellorPolicy = false;
-	}
-
-	if (game.private.lock.policyPeek) {
-		game.private.lock.policyPeek = false;
-	}
-
-	if (game.private.lock.selectPlayerToExecute) {
-		game.private.lock.selectPlayerToExecute = false;
-	}
-
-	if (game.private.lock.executePlayer) {
-		game.private.lock.executePlayer = false;
-	}
-
-	if (game.private.lock.selectSpecialElection) {
-		game.private.lock.selectSpecialElection = false;
-	}
-
-	if (game.private.lock.specialElection) {
-		game.private.lock.specialElection = false;
-	}
-
-	if (game.private.lock.selectPartyMembershipInvestigate) {
-		game.private.lock.selectPartyMembershipInvestigate = false;
-	}
-
-	if (game.private.lock.investigateLoyalty) {
-		game.private.lock.investigateLoyalty = false;
-	}
-
-	if (game.private.lock.selectPolicies) {
-		game.private.lock.selectPolicies = false;
-	}
-
-	game.gameState.pendingChancellorIndex = null;
-
-	game.private.summary = game.private.summary.updateLog({
-		enactedPolicy: team
-	});
-
-	game.general.status = 'A policy is being enacted.';
-	game.trackState[`${team}PolicyCount`]++;
-	sendGameList();
-
-	game.trackState.enactedPolicies.push({
-		position: 'middle',
-		cardBack: team,
-		isFlipped: false
-	});
-
-	sendInProgressGameUpdate(game, true);
-
-	setTimeout(() => {
-		game.trackState.enactedPolicies[index].isFlipped = true;
-		game.gameState.audioCue = team === 'liberal' ? 'enactPolicyL' : 'enactPolicyF';
-		sendInProgressGameUpdate(game, true);
-	}, process.env.NODE_ENV === 'development' ? 100 : experiencedMode ? 300 : 2000);
-
-	setTimeout(() => {
-		game.gameState.audioCue = '';
-		const chat = {
-			timestamp: new Date(),
-			gameChat: true,
-			chat: [
-				{ text: 'A ' },
-				{
-					text: team === 'liberal' ? 'liberal' : 'fascist',
-					type: team === 'liberal' ? 'liberal' : 'fascist'
-				},
-				{
-					text: ` policy has been enacted. (${
-						team === 'liberal' ? game.trackState.liberalPolicyCount.toString() : game.trackState.fascistPolicyCount.toString()
-					}/${team === 'liberal' ? '5' : '6'})`
-				}
-			]
-		};
-		const addPreviousGovernmentStatus = () => {
-			game.publicPlayersState.forEach(player => {
-				if (player.previousGovernmentStatus) {
-					player.previousGovernmentStatus = '';
-				}
-			});
-
-			if (game.trackState.electionTrackerCount <= 2 && game.publicPlayersState.findIndex(player => player.governmentStatus === 'isChancellor') > -1) {
-				game.publicPlayersState[game.gameState.presidentIndex].previousGovernmentStatus = 'wasPresident';
-				game.publicPlayersState[game.publicPlayersState.findIndex(player => player.governmentStatus === 'isChancellor')].previousGovernmentStatus =
-					'wasChancellor';
-			}
-		};
-		const powerToEnact =
-			team === 'fascist'
-				? game.customGameSettings.enabled
-					? powerMapping[game.customGameSettings.powers[game.trackState.fascistPolicyCount - 1]]
-					: presidentPowers[game.general.type][game.trackState.fascistPolicyCount - 1]
-				: null;
-
-		game.trackState.enactedPolicies[index].position =
-			team === 'liberal' ? `liberal${game.trackState.liberalPolicyCount}` : `fascist${game.trackState.fascistPolicyCount}`;
-
-		if (!game.general.disableGamechat) {
-			game.private.seatedPlayers.forEach(player => {
-				player.gameChats.push(chat);
-			});
-
-			game.private.unSeatedGameChats.push(chat);
-		}
-
-		if (game.trackState.liberalPolicyCount === 5 || game.trackState.fascistPolicyCount === 6) {
-			game.publicPlayersState.forEach((player, i) => {
-				player.cardStatus.cardFront = 'secretrole';
-				player.cardStatus.cardBack = game.private.seatedPlayers[i].role;
-				player.cardStatus.cardDisplayed = true;
-				player.cardStatus.isFlipped = false;
-			});
-
-			sendInProgressGameUpdate(game);
-
-			game.gameState.audioCue = game.trackState.liberalPolicyCount === 5 ? 'liberalsWin' : 'fascistsWin';
-			setTimeout(() => {
-				game.publicPlayersState.forEach((player, i) => {
-					player.cardStatus.isFlipped = true;
-				});
-				game.gameState.audioCue = '';
-				if (process.env.NODE_ENV === 'development') {
-					completeGame(game, game.trackState.liberalPolicyCount === 1 ? 'liberal' : 'fascist');
-				} else {
-					completeGame(game, game.trackState.liberalPolicyCount === 5 ? 'liberal' : 'fascist');
-				}
-			}, process.env.NODE_ENV === 'development' ? 100 : 2000);
-		} else if (powerToEnact && game.trackState.electionTrackerCount <= 2) {
-			const chat = {
-				timestamp: new Date(),
-				gameChat: true,
-				chat: [{ text: powerToEnact[1] }]
-			};
-
-			const { seatedPlayers } = game.private;
-
-			if (!game.general.disableGamechat) {
-				seatedPlayers.forEach(player => {
-					player.gameChats.push(chat);
-				});
-
-				game.private.unSeatedGameChats.push(chat);
-			}
-			powerToEnact[0](game);
-			addPreviousGovernmentStatus();
-
-			if (game.general.timedMode) {
-				const { presidentIndex } = game.gameState;
-
-				game.gameState.timedModeEnabled = true;
-				game.private.timerId = setTimeout(() => {
-					if (game.gameState.timedModeEnabled) {
-						const president = seatedPlayers[presidentIndex];
-						let list = seatedPlayers.filter((player, i) => i !== presidentIndex && !seatedPlayers[i].isDead);
-
-						game.gameState.timedModeEnabled = false;
-
-						switch (powerToEnact[1]) {
-							case 'The president must examine the top 3 policies.':
-								selectPolicies({ user: president.userName }, game);
-								break;
-							case 'The president must select a player for execution.':
-								if (president.role.cardName === 'fascist') {
-									list = list.filter(player => player.role.cardName !== 'hitler');
-								}
-								selectPlayerToExecute({ user: president.userName }, game, { playerIndex: seatedPlayers.indexOf(_.shuffle(list)[0]) });
-								break;
-							case 'The president must investigate the party membership of another player.':
-								selectPartyMembershipInvestigate({ user: president.userName }, game, { playerIndex: seatedPlayers.indexOf(_.shuffle(list)[0]) });
-								break;
-							case 'The president must select a player for a special election.':
-								selectSpecialElection({ user: president.userName }, game, { playerIndex: seatedPlayers.indexOf(_.shuffle(list)[0]) });
-								break;
-						}
-					}
-				}, process.env.DEVTIMEDDELAY ? process.env.DEVTIMEDDELAY : game.general.timedMode * 1000);
-				sendInProgressGameUpdate(game);
-			}
-		} else {
-			sendInProgressGameUpdate(game);
-			addPreviousGovernmentStatus();
-			startElection(game);
-		}
-
-		game.trackState.electionTrackerCount = 0;
-	}, process.env.NODE_ENV === 'development' ? 100 : experiencedMode ? 1000 : 4000);
-};
 
 /**
  * @param {object} passport - socket authentication.
